+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Fix proposal denom to use nhash #320
 - Fix some FE formatting issues #319
 - Fix some smart contracts issues seen in test #325
+- Fix delegations logic in voting #329
 
 ### Features
 - Update asset price per unit calcs for new data object #296